refactor signature in to own private module
looking ahead to #363
