+++ src/types/simple/serialize.rs
@@ -15,7 +15,7 @@ use super::TypeRow;
 
 use super::SimpleType;
 
-use super::super::AbstractSignature;
+use super::super::signature::AbstractSignature;
 
 use crate::types::type_row::TypeRowElem;
